[FIX] pos_loyalty: fix tours to search for customers
this commit fixes the tours in pos_loyalty to search for customers
when customer is not updated in list, by clicking the "Search More"
button. one tour was fixed by changing the customer name to
"A Test Partner" so that it starts with "A" and is always
available in the list.

task-4687196

closes odoo/odoo#205286

Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>