[FIX] sale_stock: avoid recomputing delivery date on confirm
Versions
--------
- 17.0+

Steps
-----
1. Have a sale order with deliverable products & no payment terms;
2. confirm order & validate delivery;
3. create an invoice;
4. modify the delivery date;
5. save changes;
6. change product quantity of a line & confirm invoice.

Issue
-----
The delivery date got reset.

Cause
-----
The `_compute_show_delivery_date` method gets called, which triggers the
recomputation of the `_compute_delivery_date` due it the latter having
`line_ids.sale_line_ids.order_id` as its `depends`.

Due to the way how `depends` works, if any of the fields in the record
chain gets modified, the compute gets triggered. In this case, because
we modified a `line_ids` record by changing the quantity, it will
therefore recompute the delivery date, overwriting the custom value.

Solution
--------
As we only want the delivery date to be recomputed when the
`effective_date` on the order changes, we should add it to the `depends`
to trigger the compute in that scenario.

In other scenarios, e.g. modifying the move or one of its lines, we
don't want to trigger a recompute, which we can achieve by always
including `delivery_date` via `_get_protected_vals` on create/write.

opw-4996654

closes odoo/odoo#229888

X-original-commit: 818cf04f0576782fe79760f21af9c9ecdbb9ce21
Signed-off-by: Steve Van Essche <svs@odoo.com>
Signed-off-by: Levi Siuzdak <sile@odoo.com>