[FIX] pos_loyalty: fix 0 discountable
When selling some product and applying a giftcard, there could be 2
giftcard lines applied on the order one with a correct amount and the
other one with 0 as amount.

Steps to reproduce:
-------------------
* Create a gift card with 50€
* Open PoS and add a Desk Pad and a Black Drawer
> Observation: There are 2 giftcard lines applied on the order

Why the fix:
------------
Sometimes the discountable amount could get weird values that are really
close to 0 but that was not interpreted as False. To fix this we use
`floatIsZero()` instead.

Note:
----------
I will introduce a hoot test in 18.3 to cover this weird use case.

opw-4866358

closes odoo/odoo#223948

Signed-off-by: Stéphane Vanmeerhaeghe (stva) <stva@odoo.com>