[FIX] mail: composition mode doesn't send message on ENTER
Before this commit, pressing ENTER to pick a composition suggestion
from IME in a discuss channel composer would send a message with
the current content of composer.

This is definitely not the intended behavior: it should change the
composing text by the IME software but the send on ENTER press should
not happen while the text is in composing.

This commit fixes the issue by adding condition for composing text.

Task-5043396

closes odoo/odoo#224340

X-original-commit: c1db6d03d06cb4e90154b993ffc21c57273dadd6
Signed-off-by: Zelong Lin (zel) <zel@odoo.com>
Signed-off-by: Alexandre Kühn (aku) <aku@odoo.com>