[FIX] website_event: search on description
Purpose
=======
Making sure that, when typing a search term, all the results displayed
in the search bar dropdown are visible on the page when clicking Enter.

Specification
=============
Previously, the search bar dropdown was displaying all the events where
the name or description matched the search term.
However, when clicking Enter, only the events where the name matched
were displayed on the page (the ones where only the description matched were ignored).

Fixing the issue by considering the description in the event search options.

Task-5039221

closes odoo/odoo#226820

X-original-commit: f7a9129b8753798a7c2aaf623cf96c5c6c5c6dc5
Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>
Signed-off-by: Amélie Dieudonné (amdi) <amdi@odoo.com>